backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["alex <me@fomin3.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asyncio = "^3.4.3"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
